requirements.txt
# MCP server dependencies
fastmcp>=2.12.4
rapidfuzz>=3.9
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.3
pdfminer.six>=20231228
